@@ -1,7 +1,7 @@
-import { defineRule } from '../utils/define-rule'
+import { defineRule } from '../utils/define-rule.js'
 const url = 'https://nextjs.org/docs/messages/no-assign-module-variable'
 
-export = defineRule({
+export const noAssignModuleVariable = defineRule({
   meta: {
     docs: {
       description: 'Prevent assignment to the `module` variable.',
@@ -12,16 +12,18 @@ export = defineRule({
     schema: [],
   },
 
-  create(context) {
+  create(context: any) {
     return {
-      VariableDeclaration(node) {
+      VariableDeclaration(node: any) {
         // Checks node.declarations array for variable with id.name of `module`
-        const moduleVariableFound = node.declarations.some((declaration) => {
-          if ('name' in declaration.id) {
-            return declaration.id.name === 'module'
+        const moduleVariableFound = node.declarations.some(
+          (declaration: any) => {
+            if ('name' in declaration.id) {
+              return declaration.id.name === 'module'
+            }
+            return false
           }
-          return false
-        })
+        )
 
         // Return early if no `module` variable is found
         if (!moduleVariableFound) {
