@@ -1,3 +1,3 @@
 import type { Rule } from 'eslint'
 
-export const defineRule = (rule: Rule.RuleModule) => rule
+export const defineRule = (rule: Rule.RuleModule): Rule.RuleModule => rule
